# src/Models/Order.cs
/*
 * Petstore Demo API
 *
 * This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/). 
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Order
    { 
 
        [DataMember(Name="id")]
        public long Id { get; set; }

        [DataMember(Name="petId")]
        public long PetId { get; set; }

        [DataMember(Name="quantity")]
        public int? Quantity { get; set; }

        [DataMember(Name="shipDate")]
        public DateTime ShipDate { get; set; }

        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum OrderStatusEnum
        {
            /// <summary>
            /// Enum PlacedEnum for placed
            /// </summary>
            [EnumMember(Value = "placed")]
            PlacedEnum = 0,
            /// <summary>
            /// Enum ApprovedEnum for approved
            /// </summary>
            [EnumMember(Value = "approved")]
            ApprovedEnum = 1,
            /// <summary>
            /// Enum DeliveredEnum for delivered
            /// </summary>
            [EnumMember(Value = "delivered")]
            DeliveredEnum = 2        }

        [DataMember(Name="status")]
        public OrderStatusEnum Status { get; set; }

        [DataMember(Name="complete")]
        public bool Complete { get; set; }

    }
}